requirements.txt
flask
waitress
numpy
pillow
torchvision